fix: use bigint FK columns for user tags
Discourse core builds bigint primary keys when migrating a fresh test
database, and CI bumps bigint sequences past the 4-byte integer range, so
proposer_id, target_user_id, tag_id, and user_id must be bigint to store
real user and tag identifiers.

+++ db/migrate/20260811120000_create_user_tags.rb
@@ -9,8 +9,8 @@ class CreateUserTags < ActiveRecord::Migration[7.0]
                null: false
 
     create_table :where_is_my_friends_user_tags do |t|
-      t.integer :proposer_id, null: false
-      t.integer :target_user_id, null: false
+      t.bigint :proposer_id, null: false
+      t.bigint :target_user_id, null: false
       t.string :label, null: false
       t.string :status, null: false, default: "pending"
       t.datetime :responded_at
@@ -35,8 +35,8 @@ class CreateUserTags < ActiveRecord::Migration[7.0]
                     on_delete: :cascade
 
     create_table :where_is_my_friends_tag_endorsements do |t|
-      t.integer :tag_id, null: false
-      t.integer :user_id, null: false
+      t.bigint :tag_id, null: false
+      t.bigint :user_id, null: false
       t.timestamps
     end
